+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added proper IMAP STARTTLS connect timeout
 - Fixed inefficient serilization for inline images and attachments
 - Added inline image and attachment size limit
+- Fixed inline images on replies and forwards
 
 
 **v0.1.38 - 03-22-2026**